requirements.txt
# Mission 8 - Technical Watch: PanCAN POC
# Python 3.10+ compatible

# ============================================
# Core Data Science
# ============================================
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ============================================
# Deep Learning - PyTorch
# ============================================
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0  # PyTorch Image Models (ResNet, ConvNeXt backbones)

# ============================================
# Transformers - Hugging Face
# ============================================
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.28.0

# ============================================
# TensorFlow/Keras (for VGG16 baseline comparison)
# ============================================
tensorflow>=2.15.0
keras>=3.0.0

# ============================================
# Machine Learning
# ============================================
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# ============================================
# Visualization
# ============================================
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# ============================================
# Image Processing
# ============================================
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0  # Data augmentation

# ============================================
# Interpretability
# ============================================
captum>=0.6.0  # PyTorch interpretability (attention visualization)
shap>=0.44.0

# ============================================
# Experiment Tracking
# ============================================
mlflow>=2.10.0
tensorboard>=2.15.0

# ============================================
# Jupyter & Development
# ============================================
jupyterlab>=4.0.0
notebook>=7.0.0
ipywidgets>=8.1.0
tqdm>=4.66.0

# ============================================
# Utilities
# ============================================
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
